feat!: use short name for deletion
