+++ cmd/oras/manifest/remove.go
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package blob
+package manifest
 
 import (
 	"errors"
@@ -22,10 +22,11 @@ import (
 
 	"github.com/spf13/cobra"
 	"oras.land/oras-go/v2/errdef"
+	oerrors "oras.land/oras/cmd/oras/internal/errors"
 	"oras.land/oras/cmd/oras/internal/option"
 )
 
-type deleteBlobOptions struct {
+type deleteOptions struct {
 	option.Common
 	option.Confirmation
 	option.Descriptor
@@ -36,23 +37,27 @@ type deleteBlobOptions struct {
 }
 
 func deleteCmd() *cobra.Command {
-	var opts deleteBlobOptions
+	var opts deleteOptions
 	cmd := &cobra.Command{
-		Use:   "delete [flags] <name>@<digest>",
-		Short: "[Preview] Delete a blob from a remote registry",
-		Long: `[Preview] Delete a blob from a remote registry
+		Use:     "rm [flags] <name>{:<tag>|@<digest>}",
+		Short:   "[Preview] Delete a manifest from remote registry",
+		Aliases: []string{"remove", "delete"},
+		Long: `[Preview] Delete a manifest from remote registry
 
 ** This command is in preview and under development. **
 
-Example - Delete a blob:
-  oras blob delete localhost:5000/hello@sha256:9a201d228ebd966211f7d1131be19f152be428bd373a92071c71d8deaf83b3e5
+Example - Delete a manifest tagged with 'latest' from repository 'localhost:5000/hello':
+  oras manifest rm localhost:5000/hello:latest
 
-Example - Delete a blob without prompting confirmation:
-  oras blob delete --yes localhost:5000/hello@sha256:9a201d228ebd966211f7d1131be19f152be428bd373a92071c71d8deaf83b3e5
+Example - Delete a manifest without prompting confirmation:
+  oras manifest rm --force localhost:5000/hello:latest
 
-Example - Delete a blob and print its descriptor:
-  oras blob delete --descriptor --yes localhost:5000/hello@sha256:9a201d228ebd966211f7d1131be19f152be428bd373a92071c71d8deaf83b3e5
-  `,
+Example - Delete a manifest and print its descriptor:
+  oras manifest rm --descriptor localhost:5000/hello:latest
+
+Example - Delete a manifest by digest 'sha256:99e4703fbf30916f549cd6bfa9cdbab614b5392fbe64fdee971359a77073cdf9' from repository 'localhost:5000/hello':
+  oras manifest rm localhost:5000/hello@sha:99e4703fbf30916f549cd6bfa9cdbab614b5392fbe64fdee971359a77073cdf9
+`,
 		Args: cobra.ExactArgs(1),
 		PreRunE: func(cmd *cobra.Command, args []string) error {
 			if opts.OutputDescriptor && !opts.Confirmed {
@@ -60,9 +65,9 @@ Example - Delete a blob and print its descriptor:
 			}
 			return opts.ReadPassword()
 		},
-		RunE: func(cmd *cobra.Command, args []string) error {
+		RunE: func(_ *cobra.Command, args []string) error {
 			opts.targetRef = args[0]
-			return deleteBlob(opts)
+			return deleteManifest(opts)
 		},
 	}
 
@@ -70,27 +75,27 @@ Example - Delete a blob and print its descriptor:
 	return cmd
 }
 
-func deleteBlob(opts deleteBlobOptions) (err error) {
+func deleteManifest(opts deleteOptions) error {
 	ctx, _ := opts.SetLoggerLevel()
 	repo, err := opts.NewRepository(opts.targetRef, opts.Common)
 	if err != nil {
 		return err
 	}
 
-	if _, err = repo.Reference.Digest(); err != nil {
-		return fmt.Errorf("%s: blob reference must be of the form <name@digest>", opts.targetRef)
+	if repo.Reference.Reference == "" {
+		return oerrors.NewErrInvalidReference(repo.Reference)
 	}
 
-	blobs := repo.Blobs()
-	desc, err := blobs.Resolve(ctx, opts.targetRef)
+	manifests := repo.Manifests()
+	desc, err := manifests.Resolve(ctx, opts.targetRef)
 	if err != nil {
 		if errors.Is(err, errdef.ErrNotFound) {
-			return fmt.Errorf("%s: the specified blob does not exist", opts.targetRef)
+			return fmt.Errorf("%s: the specified manifest does not exist", opts.targetRef)
 		}
 		return err
 	}
 
-	prompt := fmt.Sprintf("Are you sure you want to delete the blob %q?", desc.Digest)
+	prompt := fmt.Sprintf("Are you sure you want to remove the manifest %q and all tags associated with it?", desc.Digest)
 	confirmed, err := opts.AskForConfirmation(os.Stdin, prompt)
 	if err != nil {
 		return err
@@ -99,8 +104,8 @@ func deleteBlob(opts deleteBlobOptions) (err error) {
 		return nil
 	}
 
-	if err = blobs.Delete(ctx, desc); err != nil {
-		return fmt.Errorf("failed to delete %s: %w", opts.targetRef, err)
+	if err = manifests.Delete(ctx, desc); err != nil {
+		return fmt.Errorf("failed to remove %s: %w", opts.targetRef, err)
 	}
 
 	if opts.OutputDescriptor {
@@ -111,7 +116,7 @@ func deleteBlob(opts deleteBlobOptions) (err error) {
 		return opts.Output(os.Stdout, descJSON)
 	}
 
-	fmt.Println("Deleted", opts.targetRef)
+	fmt.Println("Removed", opts.targetRef)
 
 	return nil
 }
